solved lifetime dependency mismatch in the DI pipeline
